Update README.md
I utilized the OpenAI Gym to create the Mountain Car MDP environment, employing the deep SARSA technique. The environment presents two variations: discrete and continuous. For our implementation, we opted for the discrete version. Within this environment, a car is positioned randomly at the base of a valley with a sinusoidal shape. The car's only means of propulsion is through acceleration, which can be applied in any direction using a predefined set of actions. These actions are deterministically defined as three possibilities: accelerating to the left, not accelerating, or accelerating to the right.